# specs/master/contracts/todo-api-openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: API for managing user tasks in the Todo application
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server
paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
    post:
      summary: Create a new task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
        '404':
          description: Task not found
    put:
      summary: Update a task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
        '404':
          description: Task not found
    delete:
      summary: Delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
        '404':
          description: Task not found
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          description: Bearer token for authentication
          schema:
            type: string
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT
        '403':
          description: Forbidden - User ID in URL doesn't match JWT
        '404':
          description: Task not found
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the task
        user_id:
          type: string
          description: ID of the user who owns the task
        title:
          type: string
          description: Task title
        description:
          type: string
          description: Optional task description
        completed:
          type: boolean
          description: Whether the task is completed
        created_at:
          type: string
          format: date-time
          description: When the task was created
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title (required)
        description:
          type: string
          description: Optional task description
      required:
        - title
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Updated task title
        description:
          type: string
          description: Updated task description
        completed:
          type: boolean
          description: Updated completion status